refactor and clean out unused serialization classes
